+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/search/CommerceInventoryWarehouseIndexer.java
@@ -216,7 +216,7 @@ public class CommerceInventoryWarehouseIndexer
 		return super.isUseSearchResultPermissionFilter(searchContext);
 	}
 
-	private int _getItemsQuantity(
+	private BigDecimal _getItemsQuantity(
 		CommerceInventoryWarehouse commerceInventoryWarehouse) {
 
 		BigDecimal count = BigDecimal.ZERO;
@@ -233,7 +233,7 @@ public class CommerceInventoryWarehouseIndexer
 			}
 		}
 
-		return count.intValue();
+		return count;
 	}
 
 	private void _reindexCommerceInventoryWarehouses(long companyId)
