+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-delivery-catalog-impl/src/main/java/com/liferay/headless/commerce/delivery/catalog/internal/dto/v1_0/converter/ProductOptionValueDTOConverter.java
@@ -246,8 +246,7 @@ public class ProductOptionValueDTOConverter
 			_cpDefinitionInventoryEngineRegistry.getCPDefinitionInventoryEngine(
 				cpDefinitionInventory);
 
-		return BigDecimal.valueOf(
-			cpDefinitionInventoryEngine.getMinOrderQuantity(cpInstance));
+		return cpDefinitionInventoryEngine.getMinOrderQuantity(cpInstance);
 	}
 
 	private boolean _updateJSONArray(
